% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{estimate.gam.spec}
\alias{estimate.gam.spec}
\alias{estimate}
\title{Model Estimation}
\usage{
\method{estimate}{gam.spec}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.spec}.}

\item{...}{additional arguments passed to the gam function of the mgcv package.}
}
\value{
An object of class \dQuote{gam.estimate}
}
\description{
Estimates a model given a specification object.
}
\details{
The object class retains the output model from mgcv in the model slot
of the returned object list so can be dispatched to other methods created for
mgcv gam objects.
}
