% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/mgam.R
\name{fitted.gam.estimate}
\alias{fitted.gam.estimate}
\alias{fitted}
\alias{fitted.mgam.estimate}
\title{Model Fitted Values}
\usage{
\method{fitted}{gam.estimate}(object, ...)

\method{fitted}{mgam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{mgam.estimate}.}

\item{...}{not currently used.}
}
\value{
An xts vector

A data.table with the index_id, series_id and response fitted values.
}
\description{
Extract the fitted values from an estimated model.

Extract the fitted values of the estimated model.
}
