% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{gam_modelspec}
\alias{gam_modelspec}
\title{GAM Model specification}
\usage{
gam_modelspec(formula, data, family = gaussian(link = "identity"), ...)
}
\arguments{
\item{formula}{a valid gam formula.}

\item{data}{an xts matrix of the data.}

\item{family}{a valid gam distribution family. Currently on the \dQuote{gaussian}
and \dQuote{gaulss} are fully supported. In the later case, the formula is a list
of length 2 with a formula for the mean response and one for the standard deviation.}

\item{...}{additional arguments (none currently supported).}
}
\value{
An object of class \dQuote{gam.spec}.
}
\description{
Sets up a gam model via formula.
}
\note{
This is a wrapper to the gam function from the mgcv package.
}
