% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{gam_trainspec}
\alias{gam_trainspec}
\title{GAM Model Training Data Setup}
\usage{
gam_trainspec(
  formula,
  family = gaussian(link = "identity"),
  data,
  estimation_dates,
  prediction_dates,
  validate = TRUE
)
}
\arguments{
\item{formula}{a valid gam formula.}

\item{family}{a valid gam distribution family}

\item{data}{an xts data matrix}

\item{estimation_dates}{a vector of estimation dates}

\item{prediction_dates}{a list of length(estimation_dates) with the
dates to use for forecasting after each estimation date.}

\item{validate}{whether to perform validity tests on the prediction_dates
so that they are never less than or equal to estimation_dates, and that both
estimation_dates and prediction_dates actually exist in the supplied data. The
first one will generate a wwrning, and the last 2 an error.}
}
\value{
an object of class \dQuote{gam.trainspec}
}
\description{
Sets up a gam data object for backtesting.
}
