% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE113957_list}
\alias{GSE113957_list}
\title{GSE113957 example dataset list}
\format{
A list.
}
\source{
Derived from GEO: GSE113957 (pre-processed by the authors).
}
\usage{
GSE113957_list
}
\description{
A pre-processed list object used in examples and testing.
}
\keyword{datasets}
