% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgam.R
\name{mgam_modelspec}
\alias{mgam_modelspec}
\title{Multivariate GAM Model specification}
\usage{
mgam_modelspec(
  formula,
  data,
  series_id = NULL,
  index_id = NULL,
  family = gaussian(link = "identity"),
  pooled = FALSE,
  ...
)
}
\arguments{
\item{formula}{a valid gam formula.}

\item{data}{a data.table object in long format of the data, with no missing values.}

\item{series_id}{the column name which corresponds to the unique identifier
of the series.}

\item{index_id}{the column name which corresponds to the unique date/time
index of the series.}

\item{family}{a valid gam distribution family.}

\item{pooled}{whether to estimate a pooled model.}

\item{...}{additional arguments (none currently supported).}
}
\value{
An object of class \dQuote{gam.spec}.
}
\description{
Sets up an mgam model via formula.
}
\details{
The multivariate GAM model described here is at a minimum a wrapper for multiple
estimation and prediction of a set of series in a convenient wrapper, with the
caveat that they share a common formula. However, it also allows pooled estimation
as well as forecasts with coupled uncertainty via the use of a copula.
}
