% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTBLS1991_list}
\alias{MTBLS1991_list}
\title{MTBLS1991 example dataset list}
\format{
A list.
}
\source{
Derived from MetaboLights: MTBLS1991 (pre-processed by the authors).
}
\usage{
MTBLS1991_list
}
\description{
A pre-processed list object used in examples and testing.
}
\keyword{datasets}
