% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_list}
\alias{multi_list}
\title{Multi-omics example list}
\format{
A list.
}
\source{
Derived from GEO: GSE85358 and its publication (pre-processed
by the authors).
}
\usage{
multi_list
}
\description{
A combined example list object used for demonstration.
}
\keyword{datasets}
