% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_feature_young_old_gam}
\alias{plot_feature_young_old_gam}
\title{############################################################################
Two-segment generalized additive models for feature screening}
\usage{
plot_feature_young_old_gam(
  y,
  Age,
  Group,
  k = 6,
  bs = "cs",
  ci_level = 0.95,
  n_grid = 200,
  show_lm = TRUE,
  main = "Feature"
)
}
\arguments{
\item{y}{A numeric vector variables of the feature exxpression.}

\item{Age}{A numeric vector variables of length \eqn{n} containing age
information for each sample.}

\item{Group}{A binary classification vector variables of length \eqn{n}
containing age segment for each sample.}

\item{k}{An integer specifying the basis dimension of spline smoothers
used in GAM fitting.}

\item{bs}{a two letter character string indicating the (penalized) smoothing
basis to use.}

\item{ci_level}{confidence interval.}

\item{n_grid}{the number of grids in the diagram.}

\item{show_lm}{Logical, set to \code{TRUE}, if you want to show the linear
line in the diagram.}

\item{main}{the title of the diagram.}
}
\description{
This function utilizes two-segment generalized additive models to identify
variables showing significant differences between young and old segments
based on age information.
}
\examples{
## Not run:
## First example
sim <- simData(N = 40,
               mean.pat = 2,
               sd.pat = 1,
               mean.diff = 2,
               sd.diff = 1,
               mean.bg = 0,
               sd.bg = 0.5,
               p_total = 2000,
               p_pattern_relevant = 40,
               p_diff_relevant = 40,
               percentage_overlap_variables = 0.5,
               seeds = 123)

plot_feature_young_old_gam(sim$data[, 5], sim$Age, sim$Group,
k = 6, show_lm = TRUE)

}
