% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{predict.gam.estimate}
\alias{predict.gam.estimate}
\alias{predict}
\title{Model Prediction}
\usage{
\method{predict}{gam.estimate}(
  object,
  newdata,
  nsim = 9000,
  tabular = FALSE,
  distribution = NULL,
  innov = NULL,
  innov_type = "q",
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{newdata}{an xts matrix of external regressors in the forecast horizon.}

\item{nsim}{The number of simulations to use for generating the simulated predictive distribution.}

\item{tabular}{whether to return the data in long format as a data.table instead
of an object of tsmodels.predict}

\item{distribution}{A valid distribution from the tsdistributions package to be used to fit the
response residuals and then proxy for the predictive distribution. Only valid if the estimated
model used the gaussian family, else will throw an error.}

\item{innov}{an optional matrix of innovations (see innov_type for admissible types),
of dimensions nsim x horizon.}

\item{innov_type}{if \\sQuote{innov} is not NULL, then this denotes the type of values
passed, with \\dQuote{q} denoting quantile probabilities (default and
backwards compatible), \\dQuote{z} for standardized errors and \\dQuote{r} for
raw innovations (non standardized). See details.}

\item{...}{not currently used.}
}
\value{
An object which inherits class \dQuote{tsmodel.predict} with slots for
the simulated predictive distribution, the original series (as a
zoo object), the original specification object and the mean forecast. If tabular
is TRUE, will return a data.table object of the predictions in long format with
some additional columns.
Using a gaussian distribution for estimation followed by a custom distribution for fitting
the model residuals in the prediction step to proxy for the uncertainty is a 2 step option
which may yield superior results to using something like family scat in the mgcv model as this
has proven to be problematic in certain cases.
}
\description{
Prediction function for class \dQuote{gam.estimate}.
}
\details{
The \dQuote{distribution} and \dQuote{innovations} arguments allow for significant customization
of the simulated distribution. When \dQuote{innov_type} is type q, then the values
are transformed back to normally distributed predictions using the normal quantile function
with mean the prediction vector and standard deviation calculated from the model residuals.
If distribution is also not NULL, then instead of using the normal quantile function, a model
if first fitted to the residuals given the distribution chosen and the quantile function of
that distribution is used to transform the uniform value.
When \dQuote{innov_type} is either z or r, then the values are simply rescaled using the
standard deviation of the residuals (for type z), and re-centered by the prediction means
(for type z and r).
}
