% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/mgam.R
\name{residuals.gam.estimate}
\alias{residuals.gam.estimate}
\alias{residuals}
\alias{residuals.mgam.estimate}
\title{Model Residuals}
\usage{
\method{residuals}{gam.estimate}(object, type = "response", ...)

\method{residuals}{mgam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{mgam.estimate}.}

\item{type}{the of residuals with \dQuote{response} being the default. Other
options see the mgcv documentation.}

\item{...}{not currently used.}
}
\value{
An xts vector

A data.table with the index_id, series_id and response residuals values.
}
\description{
Extract the residual values from an estimated model.

Extract the residuals of the estimated model.
}
