% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgam.R
\name{sigma.mgam.estimate}
\alias{sigma.mgam.estimate}
\alias{sigma}
\title{Extract Model Sigma}
\usage{
\method{sigma}{mgam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{mgam.estimate}.}

\item{...}{not currently used.}
}
\value{
A data.table with the index_id, series_id and standard deviation (which
can be a vector if the family is \dQuote{gaulss}).
}
\description{
Extract the standard deviation of the estimated model.
}
