% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simData}
\alias{simData}
\title{Gaussian distribution Data Simulation}
\usage{
simData(
  N = 40,
  mean.pat = 2,
  sd.pat = 1,
  mean.diff = 0,
  sd.diff = 1,
  mean.bg = 0,
  sd.bg = 0.5,
  p_total = 500,
  p_pattern_relevant = 60,
  p_diff_relevant = 100,
  percentage_overlap_variables = 0.5,
  seeds = 123
)
}
\arguments{
\item{N}{Integer, the number of samples.}

\item{mean.pat}{Numeric, the mean of different pattern effect coefficients
for all related variables.}

\item{sd.pat}{Numeric, the variability (standard deviation) of different
pattern effect coefficients for all related variables.}

\item{mean.diff}{Numeric, the mean of different expressed effect coefficients
for all related variables}

\item{sd.diff}{Numeric, the variability (standard deviation) of different
expressed effect coefficients for all related variables.}

\item{mean.bg}{Numeric, the mean of background noise.}

\item{sd.bg}{Numeric, the variability (standard deviation) of
background noise.}

\item{p_total}{Integer, the number of variables.}

\item{p_pattern_relevant}{Integer, the number of variables affected by
different pattern effects.}

\item{p_diff_relevant}{Integer, the number of variables affected by
different expressed effects.}

\item{percentage_overlap_variables}{Numeric, in the range of \eqn{0}
to \eqn{1}, the percentage of variables with both confounder and treatment
effects given fixed number of variables with each effect.}

\item{seeds}{Integer. Default value is randomly selected from
\eqn{0} to \eqn{999}.}
}
\value{
\code{simData} returns a list that contains the following components:

\item{data}{The simulated data.}
\item{Age}{The simulated outcome variable indicating the Age information.}
\item{Group}{The simulated outcome variable indicating the Age group
information.}
\item{true.pat}{The variables assigned with the different pattern effect.}
\item{true.diff}{The variables assigned with the different expressed effect.}
}
\description{
This function simulates data with margin- and nonlinear-dependent age effects
from Gaussian distribution for single-omics data.
}
\examples{
sim <- simData(N = 40,
               mean.pat = 2,
               sd.pat = 1,
               mean.diff = 2,
               sd.diff = 1,
               mean.bg = 0,
               sd.bg = 0.5,
               p_total = 2000,
               p_pattern_relevant = 40,
               p_diff_relevant = 40,
               percentage_overlap_variables = 0.5,
               seeds = 123)


}
\author{
Jianzhong Lu
}
