% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gam.estimate}
\alias{summary.gam.estimate}
\alias{summary}
\title{Model Estimation Summary}
\usage{
\method{summary}{gam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{...}{additional arguments passed through to the mgcv gam summary method.}
}
\value{
A list of summary information (see mgcv summary method).
}
\description{
Summary method for class \dQuote{gam.estimate}
}
