% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{tsbacktest.gam.trainspec}
\alias{tsbacktest.gam.trainspec}
\alias{tsbacktest}
\title{Walk Forward Model Backtest}
\usage{
\method{tsbacktest}{gam.trainspec}(
  object,
  alpha = NULL,
  trace = FALSE,
  nsim = 5000,
  distribution = NULL,
  tabular = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{gam.train}.}

\item{alpha}{optional numeric vector of coverage rates for which to calculate
the quantiles.}

\item{trace}{whether to show the progress bar. The user is expected to have
set up appropriate handlers for this using the \dQuote{progressr} package.}

\item{nsim}{The number of simulations to use for generating the simulated predictive distribution.}

\item{distribution}{A valid distribution from the tsdistributions package to be used to fit the
response residuals and then proxy for the predictive distribution. Only valid if the estimated
model used the gaussian family, else will throw an error.}

\item{tabular}{whether to return the full simulated distributiion in long format
as a data.table instead of summarized distribution with quantiles.}

\item{...}{additional arguments passed to the \dQuote{gam} function during estimation.}
}
\value{
A list with the following data.tables:
\itemize{
\item prediction : the backtest table with forecasts and actuals
\item metrics: a summary performance table showing metrics by
forecast horizon (MAPE, MSLRE, BIAS and MIS if alpha was not NULL).
}
}
\description{
Generates an expanding window walk forward backtest.
}
\note{
The function can use parallel functionality as long as the user has
set up a \code{\link[future]{plan}} using the future package.
}
