% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgam.R
\name{tsconvert.mgam.estimate}
\alias{tsconvert.mgam.estimate}
\alias{tsconvert}
\title{Convert an mgam.estimate model to a list of gam.estimate models}
\usage{
\method{tsconvert}{mgam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{mgam.estimate}.}

\item{...}{not currently used.}
}
\value{
A list with each slot of class \dQuote{gam.estimate} holding
the estimated model for each series.
}
\description{
Converts an mgam.estimate model to a list of gam.estimate models
}
