% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsdecompose.gam.estimate}
\alias{tsdecompose.gam.estimate}
\alias{tsdecompose}
\alias{tsdecompose.gam.predict}
\title{Model Decomposition}
\usage{
\method{tsdecompose}{gam.estimate}(object, ...)

\method{tsdecompose}{gam.predict}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate} or \dQuote{gam.predict}}

\item{...}{not currently used.}
}
\value{
An xts matrix of the decomposed additive components.
}
\description{
Decomposes the estimated model or prediction into its component
parts (aggregated basis terms).
}
