% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgam-package.R
\docType{package}
\name{tsgam-package}
\alias{tsgam}
\alias{tsgam-package}
\title{tsgam: Interface to mgcv}
\description{
Wrappers for GAM models in the mgcv package.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

}
\keyword{internal}
