% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSGAM.R
\name{TSGAM}
\alias{TSGAM}
\title{Two-segment generalized additive models for feature screening}
\usage{
TSGAM(
  X,
  Age,
  ncomp = 1,
  keep_rate = 0.8,
  young_cut = 36,
  old_cut = 49,
  p_delta_thr = 0.05,
  adjr2_min = 0.3,
  dev_min = 0.3,
  edf_thr = 1,
  feat_k = 6,
  feat_gamma = 1.2,
  feat_nboot = 20,
  multi = FALSE,
  WGCNA = TRUE
)
}
\arguments{
\item{X}{A numeric matrix dimension \eqn{n \times p} representing features
measured on samples, or a list including such matrices
which multi-omics profiles measured on the same set of samples.}

\item{Age}{A numeric vector variables of length \eqn{n} containing age
information for each sample.}

\item{ncomp}{Integer, the number of components to include in the DIABLO
model.}

\item{keep_rate}{Numeric, The proportion of features retained by splsda.}

\item{young_cut}{A numeric value specifying the upper age threshold for the
Young segment. Samples with \code{Age <= young_cut} are assigned to the Young
segment.}

\item{old_cut}{A numeric value specifying the lower age threshold for the Old
segment. Samples with \code{Age >= old_cut} are assigned to the Old segment.}

\item{p_delta_thr}{A numeric value specifying the significance threshold
for delta.}

\item{adjr2_min}{A numeric value specifying the minimum adjusted
\eqn{R^2} required.}

\item{dev_min}{A numeric value specifying the minimum proportion of
deviance explained required.}

\item{edf_thr}{A numeric value specifying the threshold for linear or
nonlinear.}

\item{feat_k}{An integer specifying the basis dimension of spline smoothers
used in feature-level TS-GAM fitting.}

\item{feat_gamma}{A numeric value controlling the smoothness penalty
inflation in TS-GAM fitting, used to stabilize model estimation.}

\item{feat_nboot}{An integer specifying the number of bootstrap resamples
used to assess the uncertainty and significance of segment-specific effects.}

\item{multi}{Logical, should be set to \code{TRUE}, if the \code{X} input is
a list, the DIABLO is using before the TS-GAM.}

\item{WGCNA}{Logical, should be set to \code{TRUE}, if the one of \code{X}
matrics have high dimension. The WGCNA is using before the feature-level
TS-GAM.}
}
\value{
\code{TSGAM} returns a list that contains the following components:

\item{res_boot}{A long-format data frame containing bootstrap-refined TS-GAM
results for selected candidate features, including bootstrap-based standard
errors and p-values for segment-specific effects.}
\item{SigPrimary}{A data frame of broadly significant features passing
bootstrap-based significance and minimum model fit criteria, representing a
conservative set of differential candidates.}
\item{SigMarginal}{A data frame of marginal (expression-dominated)
differential features, characterized by large segment-specific effects with
similar trajectory shapes across age segments.}
\item{SigTrajectory}{A data frame of trajectory-differential features
exhibiting segment-specific nonlinear age patterns or pronounced differences
in trajectory complexity between segments.}
}
\description{
This function utilizes two-segment generalized additive models to identify
variables showing significant differences between young and old segments
based on age information.
}
\examples{
## First example
sim <- simData(N = 40,
               mean.pat = 2,
               sd.pat = 1,
               mean.diff = 2,
               sd.diff = 1,
               mean.bg = 0,
               sd.bg = 0.5,
               p_total = 2000,
               p_pattern_relevant = 40,
               p_diff_relevant = 40,
               percentage_overlap_variables = 0.5,
               seeds = 123)

result <- TSGAM(sim$data,
                sim$Age)


}
\author{
Jianzhong Lu
}
