% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmetrics.gam.predict}
\alias{tsmetrics.gam.predict}
\alias{tsmetrics}
\alias{tsmetrics.gam.estimate}
\title{Performance Metrics}
\usage{
\method{tsmetrics}{gam.predict}(object, actual, alpha = 0.1, ...)

\method{tsmetrics}{gam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate} or \dQuote{gam.predict}}

\item{actual}{the actual data matched to the dates of the forecasts.}

\item{alpha}{the coverage level for distributional forecast metrics.}

\item{...}{not currently used.}
}
\description{
Performance metrics from an estimated or predicted gam model.
}
