% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgam.R, R/mgcv-model.R
\name{estimate.mgam.spec}
\alias{estimate.mgam.spec}
\alias{estimate}
\alias{estimate.gam.spec}
\title{Model Estimation}
\usage{
\method{estimate}{mgam.spec}(object, use_bam = FALSE, ...)

\method{estimate}{gam.spec}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.spec}.}

\item{use_bam}{whether to use \code{\link[mgcv]{bam}} for very large datasets
(in the pooled case only).}

\item{...}{additional arguments passed to the gam function of the mgcv package.}
}
\value{
An object of class \dQuote{mgam.estimate}

An object of class \dQuote{gam.estimate}
}
\description{
Estimates a model given a specification object.

Estimates a model given a specification object.
}
\details{
If the model is pooled, the object class retains the output model from
mgcv in the model slot of the returned object list so can be dispatched to
other methods created for mgcv gam objects. In the case of a non-pooled model,
the output id a list of models, one for each series.
For the non pooled model, the function makes use of the future package for
parallel estimation across the series. However, care should be taken in this
case since mgcv already makes use of openMP so the user may need to control
the number of threads already committed.

The object class retains the output model from mgcv in the model slot
of the returned object list so can be dispatched to other methods created for
mgcv gam objects.
}
