% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.gam.estimate}
\alias{fitted.gam.estimate}
\alias{fitted}
\title{Model Fitted Values}
\usage{
\method{fitted}{gam.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{...}{not currently used.}
}
\value{
An xts vector
}
\description{
Extract the fitted values from an estimated model.
}
