% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv-model.R
\name{predict.gam.estimate}
\alias{predict.gam.estimate}
\alias{predict}
\title{Model Prediction}
\usage{
\method{predict}{gam.estimate}(
  object,
  newdata,
  nsim = 9000,
  tabular = FALSE,
  distribution = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{newdata}{an xts matrix of external regressors in the forecast horizon.}

\item{nsim}{The number of simulations to use for generating the simulated predictive distribution.}

\item{tabular}{whether to return the data in long format as a data.table instead
of an object of tsmodels.predict}

\item{distribution}{A valid distribution from the tsdistributions package to be used to fit the
response residuals and then proxy for the predictive distribution. Only valid if the estimated
model used the gaussian family, else will throw an error.}

\item{...}{not currently used.}
}
\value{
An object which inherits class \dQuote{tsmodel.predict} with slots for
the simulated predictive distribution, the original series (as a
zoo object), the original specification object and the mean forecast. If tabular
is TRUE, will return a data.table object of the predictions in long format with
some additional columns.
Using a gaussian distribution for estimation followed by a custom distribution for fitting
the model residuals in the prediction step to proxy for the uncertainty is a 2 step option
which may yield superior results to using something like family scat in the mgcv model as this
has proven to be problematic in certain cases.
}
\description{
Prediction function for class \dQuote{gam.estimate}.
}
