% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgam.R, R/mgcv-model.R
\name{predict.mgam.estimate}
\alias{predict.mgam.estimate}
\alias{predict}
\alias{predict.gam.estimate}
\title{Model Prediction}
\usage{
\method{predict}{mgam.estimate}(
  object,
  newdata = NULL,
  nsim = 5000,
  copula = TRUE,
  copula_family = c("Normal", "Student"),
  shape = NA,
  distribution = NULL,
  tabular = FALSE,
  ...
)

\method{predict}{gam.estimate}(
  object,
  newdata,
  nsim = 9000,
  tabular = FALSE,
  distribution = NULL,
  innov = NULL,
  innov_type = "q",
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{newdata}{an xts matrix of external regressors in the forecast horizon.}

\item{nsim}{The number of simulations to use for generating the simulated predictive distribution.}

\item{copula}{whether to use a copula to create correlated predictive distribution samples (not available
for pooled models). This requires that the newdata are fully aligned in terms of time indices i.e.
that the series share the same time indices.}

\item{copula_family}{the copula family to use.}

\item{shape}{the degrees of freedom in the case of the Student copula. An NA value means that
this will be estimated, but for large dimensional series it may be best to fix this since
estimation may be slow.}

\item{distribution}{A valid distribution from the tsdistributions package to be used to fit the
response residuals and then proxy for the predictive distribution. Only valid if the estimated
model used the gaussian family, else will throw an error.}

\item{tabular}{whether to return the data in long format as a data.table instead
of an object of tsmodels.predict}

\item{...}{not currently used.}

\item{innov}{an optional matrix of innovations (see innov_type for admissible types),
of dimensions nsim x horizon.}

\item{innov_type}{if \\sQuote{innov} is not NULL, then this denotes the type of values
passed, with \\dQuote{q} denoting quantile probabilities (default and
backwards compatible), \\dQuote{z} for standardized errors and \\dQuote{r} for
raw innovations (non standardized). See details.}
}
\value{
a list with the following slots:
\itemize{
\item (tabular = FALSE) prediction_list is a list of \dQuote{tsmodel.predict} objects, one for each series,
for which a plot method is available.
\item (tabular = TRUE) prediction_table is a long format data.table with columns for by series_id,
estimation_date, draw (simulation draw), forecast_date and value (the forecast value).
}

An object which inherits class \dQuote{tsmodel.predict} with slots for
the simulated predictive distribution, the original series (as a
xts object), the original specification object and the mean forecast. If tabular
is TRUE, will return a data.table object of the predictions in long format with
some additional columns.
Using a gaussian distribution for estimation followed by a custom distribution for fitting
the model residuals in the prediction step to proxy for the uncertainty is a 2 step option
which may yield superior results to using something like family scat in the mgcv model as this
has proven to be problematic in certain cases.
}
\description{
Prediction function for class \dQuote{gam.estimate}.

Prediction function for class \dQuote{gam.estimate}.
}
\details{
The \dQuote{distribution} and \dQuote{innovations} arguments allow for significant customization
of the simulated distribution. When \dQuote{innov_type} is type q, then the values
are transformed back to normally distributed predictions using the normal quantile function
with mean the prediction vector and standard deviation calculated from the model residuals.
If distribution is also not NULL, then instead of using the normal quantile function, a model
if first fitted to the residuals given the distribution chosen and the quantile function of
that distribution is used to transform the uniform value.
When \dQuote{innov_type} is either z or r, then the values are simply rescaled using the
standard deviation of the residuals (for type z), and re-centered by the prediction means
(for type z and r).
}
\note{
The newdata series_id must match exactly those that were used in the original
specification. The function will test for this and optionally reorder the newdata
so that the series_id order is the same as in the original dataset. This restriction
may be relaxed in a later release so that a pooled model can be used on entirely new
series.
}
