% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.gam.estimate}
\alias{residuals.gam.estimate}
\alias{residuals}
\title{Model Residuals}
\usage{
\method{residuals}{gam.estimate}(object, type = "response", ...)
}
\arguments{
\item{object}{an object of class \dQuote{gam.estimate}.}

\item{type}{the of residuals with \dQuote{response} being the default. Other
options see the mgcv documentation.}

\item{...}{not currently used.}
}
\value{
An xts vector
}
\description{
Extract the residual values from an estimated model.
}
